refactor: search component divided into two functions
