+++ frontend/src/components/Search-item/index.ts
@@ -1,0 +1,3 @@
+import SearchItem from './Search-item';
+
+export default SearchItem;
